+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Issue Fixes
 * [#107](https://github.com/upfrontsystems/izinto-angular/issues/107) Add map visualisation example
 * [#184](https://github.com/upfrontsystems/izinto-angular/issues/184) Add forecasts table
 * [#197](https://github.com/upfrontsystems/izinto-angular/issues/197) Add optional date selector flag to dashboard
+* [#199](https://github.com/upfrontsystems/izinto-angular/issues/199) Add image to collection and dashboard
 
 
 ## 0.6